+++ editor.planx.uk/src/theme.ts
@@ -27,18 +27,6 @@ export const FONT_WEIGHT_BOLD = "700";
 export const LINE_HEIGHT_BASE = "1.33";
 const SPACING_TIGHT = "-0.02em";
 
-declare module "@mui/material/styles" {
-  interface BreakpointOverrides {
-    xs: true;
-    sm: true;
-    md: true;
-    lg: true;
-    xl: true;
-    formWrap: true;
-    contentWrap: true;
-  }
-}
-
 const DEFAULT_PALETTE: Partial<PaletteOptions> = {
   primary: {
     main: DEFAULT_PRIMARY_COLOR,
